+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -1,0 +1,2 @@
+mod addresses_table;
+pub(super) use addresses_table::AddressesTableMigration;