+++ zcash_client_sqlite/CHANGELOG.md
@@ -29,6 +29,12 @@ and this library adheres to Rust's notion of
     column. Values for this column should be derived from the wallet's seed and
     the account number; the Sapling component of the resulting Unified Full
     Viewing Key should match the old value in the `extfvk` column.
+  - The `address` and `transparent_address` columns of the `accounts` table have
+    been removed.
+    - A new `addresses` table stores Unified Addresses, keyed on their `account`
+      and `diversifier_index`, to enable storing diversifed Unified Addresses.
+    - Transparent addresses for an account should be obtained by extracting the
+      transparent receiver of a Unified Address for the account.
   - A new non-null column, `output_pool` has been added to the `sent_notes`
     table to enable distinguishing between Sapling and transparent outputs
     (and in the future, outputs to other pools). Values for this column should